# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Socnet Plizi",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://vm1095330.hl.had.pm/api"
    },
    {
      "url": "http://vm1095330.hl.had.pm/api"
    }
  ],
...
  "paths": {
...
    "/user/notifications/mark/read": {
      "patch": {
        "tags": [
          "User"
        ],
        "description": "Пометить уведомления как прочитанные",
        "operationId": "markNotificationsAsRead",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/NotificationRes"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/user/friendship": {
      "get": {
        "tags": [
          "Friendships"
        ],
        "description": "Список моих друзей",
        "operationId": "myFriendsList",
          "parameters": [
              {
                  "name": "limit",
                  "in": "path",
                  "description": "Кол-во записей",
                  "required": false,
                  "schema": {
                      "type": "integer"
                  }
              },
              {
                  "name": "offset",
                  "in": "path",
                  "description": "Смещение",
                  "required": false,
                  "schema": {
                      "type": "integer"
                  }
              }
          ],
        "responses": {
          "200": {
            "description": "Успешно добавлен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "post": {
        "tags": [
          "Friendships"
        ],
        "description": "Отправить запрос на добавление в друзья",
        "operationId": "addToFriend",
        "requestBody": {
          "description": "ID пользователя которго требуется добавить в друзья",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешно добавлен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Запрос на добавление в друзья отправлен"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Вы уже отправляли запрос данному пользователю / Вы уже отправляли запрос данному пользователю",
            "content": {}
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/user/friendship/group": {
      "post": {
        "tags": [
          "Friendships"
        ],
        "description": "Добавить друга в группу",
        "operationId": "addFriendToGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "format": "int64"
                  },
                  "group": {
                    "type": "string",
                    "enum": [
                      "featured",
                      "family"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Вы добавили пользователя в группу",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "Вы добавили пользователя в группу"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Данный пользователь не в ваших друзьях",
            "content": {}
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
      "user/friendship/group/{group}/{userId}": {
          "delete": {
              "tags": [
                  "Friendships"
              ],
              "description": "Удаление пользователя из избранных",
              "operationId": "deleteFriendFromGroup",
              "parameters": [
                  {
                      "name": "group",
                      "in": "path",
                      "description": "Группа",
                      "required": true,
                      "schema": {
                          "type": "string",
                          "enum": [
                              "featured",
                              "family"
                          ]
                      }
                  },
                  {
                      "name": "userId",
                      "in": "path",
                      "description": "ID пользователя",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Успешно добавлен",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "object",
                                  "properties": {
                                      "data": {
                                          "type": "object",
                                          "properties": {
                                              "list": {
                                                  "type": "array",
                                                  "items": {
                                                      "$ref": "#/components/schemas/User"
                                                  }
                                              }
                                          }
                                      }
                                  }
                              }
                          }
                      }
                  },
                  "403": {
                      "description": "Not allowed",
                      "content": {}
                  }
              },
              "security": [
                  {
                      "Authorization": []
                  }
              ]
          }
      },
    "/user/friendship/group/{group}": {
      "get": {
        "tags": [
          "Friendships"
        ],
        "description": "Группированный список друзей",
        "operationId": "getFriendsFromGroup",
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "description": "Группа",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "featured",
                "family"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно добавлен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {}
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "NotificationRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "sender": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "id": {
                    "type": "number"
                  },
                  "postId": {
                    "type": "number"
                  }
                }
              },
              "body": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "example": "user.post.created"
              }
            }
          },
          "readAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          },
          "mutualFriendsCount": {
            "type": "number"
          },
          "profile": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "sex": {
                "type": "string"
              },
              "birthday": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "relationshipId": {
                "type": "number"
              },
              "userPic": {
                "type": "string"
              }
            }
          }
        },
        "xml": {
          "name": "User"
        }
      },
...
    },
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}